requirements.txt
Flask
python-docx
pandas
elasticsearch
lxml.html
